机器人/None_QQ/course_helper/pyproject.toml
[tool.poetry]
name = "course_helper"
version = "0.1.0"
description = "course_helper"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0-beta.1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"

[tool.nonebot]
plugins = [
#    'nonebot_plugin_biliav',
    'nonebot_plugin_simplemusic',
#    'nonebot_plugin_minesweeper',
#    'nonebot_plugin_charpic',
#    'nonebot_plugin_code',
#    'nonebot_plugin_remake',
    'nonebot_plugin_petpet',
#    'nonebot_plugin_logo',
    'nonebot_plugin_course',
]
plugin_dirs = ["src/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
